requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
sentence-transformers==2.7.0
torch==2.1.2
pymilvus==2.3.3
numpy==1.26.4
scikit-learn==1.3.2
openai==1.14.3
matplotlib==3.8.3
tkinter  # Usually comes with Python standard library